CLOUD-14394 - Using the new OAuth token by default with FS
## Problem

We want to use the new OAuth endpoint (`oauth/v1/token`) by default when a user logs in. The issue is that the desktop client does not support this. They will add in support for 6.1 so we need to target all versions prior to that so they can still use the old endpoint.

## Solution

We can get the desktop version using a new paradigm that folks who use our OAuth will start to adopt. Desktop client currently uses this format for their `client_id`, `desktop_client/{customization}` (example: `desktop_client/default`). We are going to now add version to that: `desktop_client/{customization}/{version}` (example: `desktop_client/default/6.1`).

There are a few benefits that version gives us on top of just this one feature. We can get analytics from this and it would be a nice to have for any future projects. We can start to add this same format for things like NxConnect, Mobile, even Cloud.