CLOUD-14554  - Bookmarks disappearing after short time
## Problem

Bookmarks are disappearing after a short amount of time if you just sit on the page and do nothing.

## Solution

This is happening because sometimes the query params are changing to include the `openGroups` param. Bookmarks page should not fail when this happens, but also we should not add the query param to the url when it's not needed. Both should be fixed, but Bookmarks might take more changes to fix.

I'm not sure exactly what this change does to be honest. I just found the area that was causing it and @czach suggested this. I tested it as much as I knew to test and it seems to work with no other issues.

(cherry picked from commit 68cf5a6d3ad28508a71d10fd5b81499179691327)