CLOUD-14620 - Bookmark search end time is not inclusive
## Problem

In multiple bookmark tickets I noticed people were trying to search for a specific minute (12:35 - 12:35) and getting no results causing confusion. I also think I would expect that if my bookmark said 12:35 as a time, and I search for that time it should return those results

## Solution

We should just always add 59 seconds to the end time so that the final minute is inclusive. I realize as I am writing this (lol) that means we are not including that final second.. Maybe I should just go ahead and do the full 60 seconds? That really just gives us 1 MS of overlap, and in the real world that is such a crazy small chance of happening. Or maybe I should do 59 seconds + 999 milliseconds? let me know please!