CLOUD-14638: Updated redirect logic for access-table
# Problem
Routing for bulked deleted user depended on channel partners being in the route.

# Solution
Used the activated route and checked if the amount of roles being removed was equal to records in the table.

# Note
Current solution works for now, but once we fix the routing users should be replaceable with `this.router.navigate(\['..'\])`