CLOUD-14699: Fixed issue where player kept reloading for live viewer
# Problem
There's a check that fetches the archive every 30 seconds. Eventually a side effect of the change is updating the posterUrl for the camera which is caused by it being a blob used for handling 2fa on pre-6.0 systems.

# Solution
In the player-js component do not use the posterUrl to trigger the initialization of the player.

# Changes
- Adjusted check for transports so that it doesn't eat all other errors.