CLOUD-14519 Bookmark Sharing should be enabled for SaaS only
## Problem

We have a feature switch for Bookmark Sharing, but we need to also limit this feature to only SaaS customers. This of course is only surface level since they can use the API, but we should still do it on our side.

## Solution

It was recommended that we use the `/rest/v3/system/cloud/saas` endpoint. This response has a few things in it, and we may need to use that in the future so I went ahead and typed the whole thing. But for now we only care about the `state` field and check that it equals `active`.

Note: One of the bigger changes is that I started adding a [Branded Type](https://egghead.io/blog/using-branded-types-in-typescript) for DirtyId and CleanId. I had been getting frustrated debugging errors when a type is Dirty but it's supposed to be Clean and vice versa. I think we should be using these Branded Types more in these scenarios so we these types of bugs can't happen. @awu showed me that article and I really like it if you wanna read it.