CLOUD-13745 - Bookmark Sharing - Only show share option when user has specific permission
## Problem

Previously we were letting all users that have access to view a bookmark also share that bookmark. This is wrong and we should be using the `ManageBookmark` permission. The backend was already using this, but I had forgot to check for it on the client

## Solution

We already have some good permission manager code that pulls this permission for us. So we just need to check that before displaying the Share button.