CLOUD-11873 - User without edit permission trying to visit edit camera page bug
## Problem

When a user without camera edit privileges tries to go to the edit camera URL our app currently shows a blank screen. This is because we have a resolver trying to get the data, but the data returns with an error.

## Solution

After learning a bit more, it seems the suggestion is to break this out into a guard that passes data to the resolver. Previously the resolver was doing checks and navigating if data doesn't load, but that isn't recommended. It was throwing errors in the console as well because it didn't like that flow.

So here I changed the existing CameraResolver to be a guard, which checks if the user has permissions to edit the cameraId (does not care if the id exists or not). If User can edit camera and data loads fine then we pass that data to the Resolver, which passes it to the component. If User can NOT edit camera then we redirect from the Guard (Angular recommends this way) to the `/cameras` page so they can get the normal 404 placeholder.

| **BEFORE** | **AFTER** |
| ------ | ------ |
|    ![image](/uploads/0b26a84d51a280a0b9486779819c9c4d/image.png)    |     ![image](/uploads/64eea540905b78a1fb462957f99f5d24/image.png)   |