CLOUD-14635 - Camera Settings redirects to Home page on request error
## Problem

When a user has ONLY Edit Camera permission and they go to the Camera settings page, after ~30 seconds the page will send them back to the home page. This happens because we make 2 requests that both return 403 Unauthorized (`rest/v1/licenseSummaries`, `ec2/cameraThumbnail`), and then both of those requests fight with each over trying to refresh the token.

![image](/uploads/0eb30a6201969aae9cd1de7bdaa70b7a/image.png)

## Solution

This is a short term solution, but we can just not make the request to get the preview image if we know the user doesn't have access to see it. We should probably do this anyway, but of course this isn't addressing the root cause.

The main issues to me are two fold:
1. We are not able to know when a 403 error shows up if it is because the token is expired or if the user just doesn't have access to see that thing. So we generically try to refresh the tokens even though there is no need to do that.
2. We should handle the errors better in our API service so that if two requests fail at the same time that they don't fight with each other for token refreshing and end up breaking things.

### Note

This only breaks on WebAdmin. I didn't look into why for now, but on Cloud we don't do all of the refresh token and retry logic.

![image](/uploads/525163302e71349c9588f20f5f325ec8/image.png)