CLOUD-14256 - Session Truncated opens a login dialog
## Problem

Session Truncated error will occur when the server has set a shorter time limit on the user's login session.

Currently, when the intercepter catches the Session Truncated error it will automatically redirect the user to the Auth page and then Auth page will send them back. This has a few problems:
* User is forced to redirect and has no option to just go to the home page (since they are still authorized to see other Systems)
* We have a hard to fix bug in our code that doesn't allow us to load our app with a `Code` query param if the user is already logged in.
* Due to the bug above, we were previously logging out the user entirely when this Session Truncated error happened. This was not an ideal behavior

## Solution

We can solve all of this if we open a dialog and let the user login immediately. Then we can take the new login code, do the exchanges necessary, and refresh the page.

The happy path flow goes like this:
1. User makes a request to a system
2. Session Truncated error is returned, meaning we need to login again if we want to access this system
3. A dialog is opened, [using an iframe that Chris initially made here](https://gitlab.nxvms.dev/dev/cloud_portal/-/merge_requests/8388), showing our Auth app
4. The Auth app finishes and inside the iframe we catch it's URL change, which contains the new Login Code
5. We use the Login Code to call `api/account/renewSession`, which makes our existing `SessionId` have an up to date login
6. We logout of all existing systems, which erases all the old system specific access tokens (since they are now out of date)
7. We reload the page. This is admittedly a crutch that we shouldn't use very often, but in this case it would be very complicated to fix.

Other flow:
User can click outside of the dialog and that will redirect them to the home page. Ideally there would be like an X button to close the dialog as well, but since it's an iframe that would be pretty complicated.

TODOS:
* I need to test this with 2FA. I think there might need to be a little tweak to it for that. Will do it in a follow up

![image](/uploads/0f16633ecda9eb1f1da4331d9441791e/image.png)