CLOUD-11927 - Camera Setting icons look weird
## Problem

On Camera Settings the View and Detailed Info button icons were a little messed up. This is because they were made to be colored using `fill` but in this MR (https://gitlab.nxvms.dev/dev/cloud_portal/-/commit/ff5850ea3eeb9556d4105f97967f4d251f40ebce) we changed all of the button icons to be stroke instead of fill.

## Solution

I didn't wanna break any of the other icons by changing it back to fill, and since this camera settings page is gonna be deprecated very soon I am just copy/pasting the mixin but changing what I need to fix this. It's not perfect, but I think this mixin and the existing camera settings will be gone soon, and this is a small bug fix that I would otherwise ignore if the fix wasn't so easy.

Here is the mixin that I took this from for reference: https://gitlab.nxvms.dev/dev/cloud_portal/-/blob/develop/front_end/common/styles/common/_mixins.scss#L133

(this pics don't do it justice since it's small, but to my eyes it looks better)
| BEFORE | AFTER |
| ------ | ------ |
|    ![image](/uploads/abc7d1950a61d4e075ca2917764b7749/image.png)    |    ![image](/uploads/0b6558e79036e7eef493b2a6c1db7cb8/image.png)    |