CLOUD-14517 - Info Block mini refactor to fix a bug
## Problem

The Info Block component was set up in a way that made things not line up properly when text wraps.

![image](/uploads/fb93462882a085a85fbb131023a78804/image.png)

This is caused by the way the component was setup, because the Keys are in a list of divs that are not connected to the list of Value divs. This makes it hard to keep them in sync. It also makes it so that you can't select a row of text which is pretty annoying as well. 

| | |
| ------ | ------ |
|    ![image](/uploads/b85ee2d9284179e303945cc0eea84cdc/image.png)    |     ![image](/uploads/f4448012835813e2ac9a32e7ef4cb7b6/image.png)    |

## Solution

Refactor the component and reduce the complexity of it all around. This may cause some tiny differences if you pixel peep, but it's well worth it to reduce the complexity of the component. Now we can use css grid and have the keys and values side by side.

| Before | After |
| ------ | ------ |
|   ![image](/uploads/8e59a882e189a9e6fb4adea4473715c5/image.png)    |    ![image](/uploads/c4e21cde9085c26a4bf92096842f8290/image.png)    |
|   ![image](/uploads/aa322e20c392f228a9ab60d44d4da5a5/image.png)    |    ![image](/uploads/c34a31b3e99f31a26c1874027ee919d8/image.png)    |
| ![image](/uploads/362a04fda6bc36b8a58f01719692708a/image.png) | ![image](/uploads/7cce90a6e25cb663a645ec41fbb945e0/image.png) |
| ![image](/uploads/835dc00b030fbeec89df67b1c21f4486/image.png) | ![image](/uploads/4a00febc117e25a944e60fc08469cd07/image.png) |
| ![image](/uploads/a620b3fb3aa834af7621b30e72b68544/image.png) | ![image](/uploads/0ae83ac02643bfef37fef03e72e2002c/image.png) |