CLOUD-14706: Fix handling for VMS General Information context being renamed on prod
# Problem
VMS General Information context is named General Info on prod and stage.

This causes the query to quest the context template not to return a match.

# Solution
Update query so that it matches context that starts with General so it works on prod and stage but also other instances.

(cherry picked from commit 109784cc64196b60611bcab797aed68fac8af44b)