CLOUD-14720: Change how users is updated for the access table.
# Problem
Cached users from usersCacheEntityMap updated for the root, but didn't for the current group so access roles wouldn't update.

# Solution
Make the accessTableRecords use the '' group for the users roles.

# Other changes
Updated the user cache entity in the org store, because it sometimes wasn't updating when adding or removing a user. As a result the UI sometimes got stuck where a user would be added and not appear, or remove and not disappear.

(cherry picked from commit 60e1cc07ed81f0dac7a4b80518ee8c5558d2243c)