CLOUD-14706: Fix API type handling on readonly apis endpoint
# Problem
Readonly API's menu cache wasn't being updt

Issue was the way that the endpoint handled when there was no type query param.

It was using a cache that doesn't match an api type so it was never being cleared.

# Solution
Added the query param to the request from the frontend for readonly_apis.

Also added a fallback to VMS on the endpoint if a type is not provided.

(cherry picked from commit 00dab30e3fdc6070ff0a09412a5882cba774cf2f)