CLOUD-14772: Added an exception handler for when backup codes are used to toggle 2fa for a system
# Problem
Backup codes can only be used once. So after the system's 2fa status is toggled the initial backup code cannot be used to verify the users current session.

# Solution
Added a try/except that eats the exception thrown for verifying the user's session with a used backup code. The only major problem is that if the user's session hasn't been verified already they will be forced to so by the frontend.

# Note
The spec needs to be changed so that users can only use backup codes for the following
- Login
- Password Reset
- Turning off 2fa

(cherry picked from commit efea7c6d9c66562f82642ee088a28ac1390f5e1f)