CLOUD-14648 - Device and Tag filter in Bookmarks updated to be Filter Component
## Problem

We previously created a more unified solution to the filters for Bookmarks. This was to help prep for advanced object search. Now we can shift Devices and Tags to these new filters as a starting point. Visible changes should be minimal. These filter will change the way they look in Cloud 2.0 as well, but hopefully that will only require a change to the filter component now instead of all the individual filters.

## Solution

Utilize the existing structure of `FilterComponent` to make Devices and Tags work.
```html
<filter-container header="Devices"> // this does the logic for the clear and adds styling
  <single-filter [(ngModel)]="selectedDevice"> // This does the logic for selecting objects, adding the show more button, and opening the dialog
    <filter-item> // These are the options available to the filter
      Device ABC
    </filter-item>
    <filter-item>
      Device XYZ
    </filter-item>
  </single-filter>
</filter-container>
```

We are able to remove a bunch of old code now with the old devices and tags filters and their dialogs.