CLOUD-14826: Fixed routing for when partner user demotes themselves
# Problem
Part of the routing was broken so users get redirected to 404. The app couldn't handle that state so it was broken until the user reloaded the page.

# Solution
When the user changed their own role made sure to update the ownRolesIds, and reverted changes from CLOUD-13193 that redirect users to reports.