CLOUD-14794 - Shared Bookmark - Change width to static amount
## Problem

Shared Bookmarks are in sort of limbo waiting for design decisions, but there is a request to get a slightly better looking version to show to partners. One small bug is the width of the window. Design asked for it to be just a static 80% of the page

## Solution

Remove the existing max width and change it to static `width: 80%`. Again, this is sort of temporary while we wait for Cloud 2.0 redesign, but just to show something slightly better for demos we can do this.