CLOUD-14854 - Bookmark Sharing - Page open for long time bug
## Problem

If you leave our app open for a long time (multiple hours) then there is a bug with the default time a bookmark is shared. This is because the Bookmark Share component was creating a value for `OneDayFromNow` outside of the component (dumb mistake from me :face_palm:)

## Solution

Move that value inside the component so that it is re-established when the component is rendered. It is only ever used in the constructor.