CLOUD-15010: Fixed issue updating store after adding user
# Problem
When a user was added to a folder sometimes all users in the org would appear in the table and some rows would be broken.

# Solution
When creating patches for adding the new user to the org swap the id from currentGroupId to groupId so that the correct groups are updated. Also after the user has been added update the selected folder for the group cache.

# Bonus
Fixes some translation issues.