CLOUD-14938 - WebAdmin showing the browser native login box
## Problem

There was a ticket opened specifically for Session Limiter showing the native login box, but this is an issue that exists outside of this specific feature.

From what I can see in testing, I believe the issue is just that we maybe forgot to add the `/web` prefex on POST calls.

## Solution

I added the same logic to the POST request. You can see in that file that it already exists on GET/PUSH/PATCH/DELETE methods. Again, I am guessing we accidentally missed it?