CLOUD-14936 - Session Limiter - Wrong text shown on login
## Problem

When a user session expires we show this new login dialog, but the text displayed is wrong.

## Solution

I was going to try and use an existing text, since I think we don't want to add too many login text options, but none of them make any sense for this use case. I think the text options are getting changed in Cloud2.0 anyway, but we need something at least kinda sensible for the meantime.

| BEFORE | AFTER |
| ------ | ------ |
|    ![image](/uploads/51a3f9addf7777790887a5e4487ab36c/image.png)    |    ![image](/uploads/340ad19dd9d78d90ea47613f9831a941/image.png)    |