CLOUD-14937 - WebAdmin - Cloud Access Token disappears when opening app in new tab
## Problem

For Session Limiter we have a banner that shows your time remaining before you need to login. This time remaining is derived from the JWT Token which is stored in the Session Storage. So when you open a new tab that Session Storage does not exist anymore.

This is resolved on Cloud Portal because we fetch that JWT and store it in Session Storage again, but on WebAdmin we don't because no one really needs it right now.

## Solution

I added a method that fetches that access token, and if it doesn't exist yet then it will make a small API call to get it and store it.

Then I changed the banner service to use the new method to grab the access token.