CLOUD-15130 - Shared Filter in Bookmarks bug using old NxRadio component
## Problem

The Radio component has this odd Tristate super position, where it's selected, but it's different than just being selected? I'm sure there is some historical reason for this, but for the Shared Filter we just want the UI of the Radio. We don't need any of the weird logic for it.

There is a separate issue right now the Tristate position is broken and the CSS variable for it is not working (`fill: var(--brand-core-t50);`). That is resolved in a different MR here: https://gitlab.nxvms.dev/dev/cloud_portal/-/merge_requests/9118

## Solution

I didn't realize I was accidentally using the Tristate position for the Sharing Filter. So in this MR I just want to AVOID the Tristate position and have it be a normal selected state.

![image](/uploads/01f252994764b44d9f7f7e592fdd5ea8/image.png)