CLOUD-14998: Fix Email Input Error
# Problem
On the Login dialog, when copying and pasting an email address after the required input error, the error remained. 

# Changes
Added conditional logic to check for whether or not the email input field is coming from the authorize component. If so, just return `null` (similar to how the previous code was working before some updates).

![Screenshot_2024-11-08_at_1.54.43_PM](/uploads/c80dec144f8eafd890780c7e8b7a5f92/Screenshot_2024-11-08_at_1.54.43_PM.png)