CLOUD-15174 CLOUD-15180 - QR Code Phase 1 test page
## Problem

We are starting a new feature called QR Code Server Deployment. We have broken this feature into 3 Phases. All of this can be found in the links at the bottom of this MR including the Phase 1-3 strategy.

This MR creates a test page for us to work off of and give Backend/VMS/QA a place to start testing things.

#### THIS WILL NEVER BE USED IN PRODUCTION

All of this will be deleted once we finish Phase 3, which will involve rewriting all of this properly. This is a strategy we want to try for this particular feature which we hope will allow us to write better production code and also deliver quick iteration for all teams involved to test with.

Since this feature involves heavy integration between Web, Backend, VMS, and even external partners (who will be printing these QR codes on their servers). We need to hurry and start testing the whole flow since this will be a critical 6.1 feature.

## Solution

For this MR it is just creating a blank page behind Auth and a Feature Switch. Nothing special and it's not supposed to look like anything in particular. Just a little bit of css for ease on the eyes while testing.

### Resources
[Figma Link](https://www.figma.com/design/r8iMzfeVYNMTwPcwNaoZyE/Server-deployment-(QR)?node-id=6266-6360&t=MQebEnzZZOOP5OGw-1)
[QR Code Specs](https://networkoptix.atlassian.net/wiki/spaces/FS/pages/3090776217/QR+code+feature#3.1.-QR-code-creation)
[Cloud Phase 1-3 Strategy documentation](https://docs.google.com/document/d/1EUBdQRsgbBxK1kR7Q2u8incpRj_7X5x3duBN53ptZVc/edit?usp=sharing)