CLOUD-15064 - Remove Unused API field from GetServerInfo
## Problem

I reached out to backend and they told me this field, that was a little annoying to deal with in types, is not being used. They said they are going to remove that field

## Solution

We also aren't using it for anything so removing it now while I am thinking about it.