CLOUD-15171: Fixed archive playback for 6.0 systems
# Problem
Position for archive was appended in a way that created malformed query params.

# Solution
Updated the getPlaybackUrl to use URLSearchParams to remove the guess work for building the media urls. Guess work in this case is the & between each param.

# Note
Also added a fix to remove ticket for the url for hls streams. Seems to be breaking playback for future versions of the mediaserver.

(cherry picked from commit 5da85049b3302d3422bf7f6ba9f4146a46d0d656)