CLOUD-15176 CLOUD-15177 - QR Code - Site picker page and redirect to enter QR Codes
## Problem

For QR Code we need a page to select a Site and then redirect the user to a page to enter their Deployment Code.

To reiterate from previous MRs. There is no focus on style or even super clean code. Currently we are trying to make an end to end test solution for everyone to try out the QR code feature and find bugs in our process.

## Solution

Utilize existing APIs and components to show all of the Sites the user has access to. Then when they click on that we send them to the next page with the SiteId in the URL.

Then the user should enter the deployment code in the text box. Will do 1 more follow up MR to make the request to QR Code service when that is fixed.

| Site Selection | Deployment Code page |
| ------ | ------ |
|    ![image](/uploads/57504683f2556027b06bb8594831016a/image.png)    |    ![image](/uploads/ee3ef22b366c5844ff6ea2ce3367a281/image.png)    |