CLOUD-15323: Unify translations config name and location
Create folder with `config.yaml` for the automated [Crowdin](https://translate.networkoptix.com)
integration. An automated Jenkins job reads the list of the translatable branches on the Crowdin
portal, then finds corresponding branches in our repositories and translates them by the config.

Branch naming in the Crowdin is agreed to be like `[cloud_portal]develop`: name of the repository
in the square brackets, name of the branch then, without spaces.