CLOUD-14917 - Relative Color Mixin and Demo page
## Problem

For our interaction states (hover, pressed, selected) we need to adjust the color of the object by a specific amount in the HSL formula. This comes from Design specs and is consistent across all themes.

![image](/uploads/9cd8ec0939318a8316d1f10b72710ca5/image.png){width=806 height=422}

## Solution

To do this, instead of generating all these colors at runtime or doing more tedious tasks like manually declaring these states, we will instead be using `Relative Color` which is a sort of new CSS API. [CanIUse Link](https://caniuse.com/?search=relative%20color)

There is one annoyance in this, which is that Safari has a bit of an older spec. We can solve this though with the `@Supports` CSS tag and this reduces the complexity of our colors dramatically. We also can hopefully remove this supports structure soon when Safari updates and we have a wide enough adoption.

This MR creates the mixins required to generate these colors and also creates a small test page to show it working across a range of colors.

![image](/uploads/d2002673c582daf33beff8b3b3e2ff53/image.png)