CLOUD-15088: Fix IPVD Table Routing Issue
# Problem

Due to a previous change to the base table component to address a sorting issue with CP and Org users tables, the routing for IPVD devices wasn't working correctly. 

# Solution
Added conditional logic to address the issue with CP and Org users tables, while preventing the issue with IPVD (and additional components) using the base table.