CLOUD-14918 - NxComponents Icon - Fix bug with Stroke and Fill colors
## Problem

On Safari and Firefox there is a bug with our new icon component that doesn't apply the fill and stroke colors properly. This is because Firefox and Safari force all attributes to be lowercase, so it wasn't getting caught by our css selector.

## Solution

Make these changes to better support these browsers:

`applyFill` -> `apply-fill`

`applyStroke` -> `apply-stroke`