CLOUD-14560, CLOUD-15455: Fix cameras stuck loading
# Problem
Sometimes requests were stalling which was timing out the process to create webrtc connection.

# Solution
Profiling showed a couple spots in the code, mostly from grabbing image from canvas that was blocking the main thread
which was causing requests to stall. Solution was to use an off screen canvas with convertToBlob instead of using
toDataURL on canvas.