CLOUD-15395: Fix Settings Menu Scroll Issue
The menu was reinitializing unnecessarily every time the system is updated (every 20 to 30 seconds based on the polling mechanism). This was causing it to scroll back to the top. Initially it appeared to be on the click of a menu item, but that was purely coincidental.

Instead of always reinitializing the menu, now it only reinitializes if the content in the menu changes, specifically when the number of elements changes, so that the height can change accordingly.

![Screen_Recording_2025-01-08_at_10.56.48_AM](/uploads/2108152f8664e4b209a156085f15cfe1/Screen_Recording_2025-01-08_at_10.56.48_AM.mov)

(cherry picked from commit 1b0f0330f34192ebfcc1a45009b8feecd78f332f)